requirements.txt
attrs==16.3.0
beautifulsoup4==4.6.0
Django==1.10.5
PyYAML==3.11
argparse==1.2.1
django-appconf==1.0.2
django-bower==5.2.0
django-compressor==2.1
django-model-utils==2.2
django-pandas==0.4.1
futures==3.0.5
gunicorn==19.1.0
newrelic==2.22.1.20
numpy==1.12.0
pandas==0.19.2
psycopg2==2.5.3
python-dateutil==2.2
pytz==2014.4
raven==5.0.0
requests==2.12.5
six==1.7.3
